requirements.txt
autopep8==1.6.0
prompt-toolkit==3.0.30
pycodestyle==2.9.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
termcolor==1.1.0
toml==0.10.2
wcwidth==0.2.5
